+++ scrapli/driver/core/cisco_iosxe/base_driver.py
@@ -26,11 +26,22 @@ PRIVS = {
     ),
     "configuration": (
         PrivilegeLevel(
-            pattern=r"^[a-z0-9.\-_@/:]{1,63}\([a-z0-9.\-@/:\+]{0,32}\)#$",
+            pattern=r"^[a-z0-9.\-_@/:]{1,63}\((?!tcl)[a-z0-9.\-@/:\+]{0,32}\)#$",
             name="configuration",
             previous_priv="privilege_exec",
             deescalate="end",
             escalate="configure terminal",
+            escalate_auth=False,
+            escalate_prompt="",
+        )
+    ),
+    "tclsh": (
+        PrivilegeLevel(
+            pattern=r"^[a-z0-9.\-_@/:]{1,63}\(tcl\)#$",
+            name="tclsh",
+            previous_priv="privilege_exec",
+            deescalate="tclquit",
+            escalate="tclsh",
             escalate_auth=False,
             escalate_prompt="",
         )